refactor(activerecord): fold _performInsert into Persistence#_createRecord
