+++ packages/activerecord/src/attribute-methods/dirty.ts
@@ -151,7 +151,6 @@ interface DirtyPrivateHost {
   /** @internal */
   _readAttribute(name: string): unknown;
   _writeAttribute(name: string, value: unknown): void;
-  _performInsert(): Promise<unknown>;
   _performUpdate(): Promise<number>;
   changedAttributeNamesToSave: string[];
   constructor: {
